+++ R/fit-model.R
@@ -8,10 +8,10 @@ prepare_spec_fit <- function(pjnz, proj.end=2016.5, popadjust = NULL, popupdate=
 
   ## epp
   eppd <- epp::read_epp_data(pjnz)
-  epp.subp <- epp::read_epp_subpops(pjnz)
-  epp.input <- epp::read_epp_input(pjnz)
+  ## epp.subp <- epp::read_epp_subpops(pjnz)
+  ## epp.input <- epp::read_epp_input(pjnz)
 
-  epp.subp.input <- epp::fnCreateEPPSubpops(epp.input, epp.subp, eppd)
+  ## epp.subp.input <- epp::fnCreateEPPSubpops(epp.input, epp.subp, eppd)
 
   country <- attr(eppd, "country")
   cc <- attr(eppd, "country_code")
@@ -51,7 +51,7 @@ prepare_spec_fit <- function(pjnz, proj.end=2016.5, popadjust = NULL, popupdate=
     mapply(function(set, value){ attributes(set)[[attrib]] <- value; set}, obj, value.lst)
 
   val <- set.list.attr(val, "eppd", eppd)
-  val <- set.list.attr(val, "eppfp", lapply(epp.subp.input, epp::fnCreateEPPFixPar, proj.end = proj.end))
+  ## val <- set.list.attr(val, "eppfp", lapply(epp.subp.input, epp::fnCreateEPPFixPar, proj.end = proj.end))
   val <- set.list.attr(val, "specfp", specfp.subp)
   val <- set.list.attr(val, "country", read_country(pjnz))
   val <- set.list.attr(val, "region", names(eppd))
@@ -246,11 +246,14 @@ prepare_national_fit <- function(pjnz, upd.path=NULL, proj.end=2013.5, hiv_steps
 
   ## aggregate census data across regions
   ancrtcens <- do.call(rbind, lapply(eppd, "[[", "ancrtcens"))
-  if(!is.null(ancrtcens) && nrow(ancrtcens)){
-    ancrtcens$x <- ancrtcens$prev * ancrtcens$n
-    ancrtcens <- aggregate(cbind(x,n) ~ year, ancrtcens, sum)
-    ancrtcens$prev <- ancrtcens$x / ancrtcens$n
-    ancrtcens <- ancrtcens[c("year", "prev", "n")]
+  if(!is.null(ancrtcens)) {
+    ancrtcens <- subset(ancrtcens, !is.na(prev) & !is.na(n))
+    if(nrow(ancrtcens)){
+      ancrtcens$x <- ancrtcens$prev * ancrtcens$n
+      ancrtcens <- aggregate(cbind(x,n) ~ year, ancrtcens, sum)
+      ancrtcens$prev <- ancrtcens$x / ancrtcens$n
+      ancrtcens <- ancrtcens[c("year", "prev", "n")]
+    }
   }
   
   attr(val, "eppd") <- list(anc.used = do.call(c, lapply(eppd, "[[", "anc.used")),
@@ -286,7 +289,7 @@ fitmod <- function(obj, ..., epp=FALSE, B0 = 1e5, B = 1e4, B.re = 3000, number_k
   eppd <- attr(obj, "eppd")
 
   has_ancrtsite <- exists("ancsitedat", eppd) && any(eppd$ancsitedat$type == "ancss")
-  has_ancrtcens <- exists("ancrtcens", eppd) && nrow(eppd$ancrtcens)
+  has_ancrtcens <- !is.null(eppd$ancrtcens) && nrow(eppd$ancrtcens)
   
   if(!has_ancrtsite)
     fp$ancrtsite.beta <- 0
@@ -306,18 +309,17 @@ fitmod <- function(obj, ..., epp=FALSE, B0 = 1e5, B = 1e4, B.re = 3000, number_k
   likdat <- prepare_likdat(eppd, fp)
   fp$ancsitedata <- as.logical(nrow(likdat$ancsite.dat$df))
 
-  if(fp$eppmod %in% c("rhybrid", "logrw", "rlogistic_rw")){  # THIS IS REALLY MESSY, NEED TO REFACTOR CODE
-
+  if(fp$eppmod %in% c("logrw", "rhybrid")) { # THIS IS REALLY MESSY, NEED TO REFACTOR CODE
+    
     fp$SIM_YEARS <- as.integer(max(likdat$ancsite.dat$df$yidx,
                                    likdat$hhs.dat$yidx,
                                    likdat$ancrtcens.dat$yidx,
                                    likdat$hhsincid.dat$idx,
                                    likdat$sibmx.dat$idx))
-
+      
     fp$proj.steps <- seq(fp$ss$proj_start+0.5, fp$ss$proj_start-1+fp$SIM_YEARS+0.5, by=1/fp$ss$hiv_steps_per_year)
   } else
     fp$SIM_YEARS <- fp$ss$PROJ_YEARS
-
 
   ## Prepare the EPP model
   tsEpidemicStart <- if(epp) fp$tsEpidemicStart else fp$ss$time_epi_start+0.5
@@ -327,12 +329,10 @@ fitmod <- function(obj, ..., epp=FALSE, B0 = 1e5, B = 1e4, B.re = 3000, number_k
     fp <- prepare_ospline_model(fp, tsEpidemicStart=tsEpidemicStart)
   else if(fp$eppmod == "rtrend")
     fp <- prepare_rtrend_model(fp)
-  else if(fp$eppmod == "rhybrid")
-    fp <- prepare_hybrid_r(fp)
   else if(fp$eppmod == "logrw")
     fp <- prepare_logrw(fp)
-  else if(fp$eppmod == "rlogistic_rw")
-    fp <- prepare_rlogistic_rw(fp)
+  else if(fp$eppmod == "rhybrid")
+    fp <- prepare_rhybrid(fp)
 
   fp$logitiota = TRUE
 
@@ -458,20 +458,20 @@ simfit.specfit <- function(fit,
     if(rwproj)
       fit <- rw_projection(fit)
 
-    fp.list <- lapply(fit$param, function(par) update(fit$fp, list=par))
+    fp_list <- lapply(fit$param, function(par) update(fit$fp, list=par))
     mod.list <- lapply(fp_list, simmod)
 
   } else {
-    fp.list <- rep(fit$fp, length(mod.list))
+    fp_list <- rep(fit$fp, length(mod.list))
   }
   
   fit$rvec <- sapply(mod.list, attr, "rvec_ts")
   fit$prev <- sapply(mod.list, prev)
-  fit$incid <- mapply(incid, mod = mod.list, fp = fp.list)
+  fit$incid <- mapply(incid, mod = mod.list, fp = fp_list)
   fit$popsize <- sapply(mod.list, colSums, dims=3)
 
   if(inherits(pregprev, "data.frame"))
-    fit$pregprev <- mapply(agepregprev, mod=mod.list, fp=fp.list,
+    fit$pregprev <- mapply(agepregprev, mod=mod.list, fp=fp_list,
                            MoreArgs=list(aidx=pregprev$aidx, yidx=pregprev$yidx, agspan=pregprev$agspan))
   else if(is.logical(pregprev) && pregprev == TRUE)
 
@@ -508,7 +508,7 @@ simfit.specfit <- function(fit,
     fit$artcov <- sapply(mod.list, artcov15plus)
 
   if(ancartcov)
-    fit$ancartcov <- mapply(agepregartcov, mod.list, fp.list, MoreArgs=list(aidx=1, yidx=1:fit$fp$ss$PROJ_YEARS, agspan=35, expand=TRUE))
+    fit$ancartcov <- mapply(agepregartcov, mod.list, fp_list, MoreArgs=list(aidx=1, yidx=1:fit$fp$ss$PROJ_YEARS, agspan=35, expand=TRUE))
 
   if(mxoutputs){
     fit$agemx <- abind::abind(lapply(mod.list, agemx), rev.along=0)