+++ R/rt-models.R
@@ -1,77 +1,3 @@
-prepare_hybrid_r <- function(fp, tsEpidemicStart=fp$ss$time_epi_start+0.5, rw_start=fp$rw_start, rw_dk=NULL){
-  
-  if(!exists("rtpenord", fp))
-    fp$rtpenord <- 2L
-  
-  if(is.null(rw_start))
-    rw_start <- max(fp$proj.steps)
-  
-  ## if(exists("knots", fp))
-  ##   fp$numKnots <- length(fp$knots) - 4
-
-  fp$tsEpidemicStart <- fp$proj.steps[which.min(abs(fp$proj.steps - tsEpidemicStart))]
-
-  rt <- list()
-  rt$spline_steps <- fp$proj.steps[fp$proj.steps >= fp$tsEpidemicStart & fp$proj.steps <= rw_start]
-  rt$rw_steps <- fp$proj.steps[fp$proj.steps > rw_start & fp$proj.steps <= max(fp$proj.steps)]
-  
-  rt$nsteps_preepi <- length(fp$proj.steps[fp$proj.steps < tsEpidemicStart])
-  
-  if(!exists("n_splines", fp))
-    n_splines <- 7
-  else
-    n_splines <- fp$n_splines
-  
-  if(!exists("n_rw", fp))
-    n_rw <- ceiling(diff(range(rt$rw_steps)))  ##
-  else
-    n_rw <- fp$n_rw
-  
-  
-  rt$n_splines <- n_splines
-  rt$n_rw <- n_rw
-  rt$n_param <- rt$n_splines+rt$n_rw
-  
-  fp$numKnots <- rt$n_splines+rt$n_rw
-  
-  if(rt$n_splines > 0){
-    rt$spline_penord <- fp$rtpenord
-    proj.dur <- diff(range(rt$spline_steps))
-    rvec.knots <- seq(min(rt$spline_steps) - 3*proj.dur/(rt$n_splines-3), max(rt$spline_steps) + 3*proj.dur/(rt$n_splines-3), proj.dur/(rt$n_splines-3))
-    
-    fp$splineX <- splines::splineDesign(rvec.knots, rt$spline_steps)
-    
-    m <- matrix(0, rt$n_splines, rt$n_splines)
-    m[,1] <- 1
-    for(i in 2:rt$n_splines)
-      m[i:rt$n_splines,i] <- 1:(rt$n_splines-i+1)
-
-    rt$splineX <- fp$splineX %*% m
-  }
-  
-  ## Random walk design matrix
-  if(!is.null(rw_dk))
-    rt$rw_knots <- seq(min(rt$rw_steps), max(rt$rw_steps)+rw_dk, by=rw_dk)
-  else 
-    rt$rw_knots <- seq(min(rt$rw_steps), max(rt$rw_steps), length.out=n_rw+1)
-  rt$rwX <- outer(rt$rw_steps, rt$rw_knots[1:n_rw], ">=")
-  class(rt$rwX) <- "integer"
-
-  rt$eppmod <- "rhybrid"
-  fp$rt <- rt
-
-  fp$rvec.spldes <- rbind(matrix(0, rt$nsteps_preepi, fp$numKnots),
-                          cbind(rt$splineX, matrix(0, length(rt$spline_steps), n_rw)),
-                          cbind(matrix(tail(rt$splineX, 1), nrow=length(rt$rw_steps), ncol=n_splines, byrow=TRUE), rt$rwX))
-                                
-  if(!exists("eppmod", fp))
-    fp$eppmod <- "rhybrid"
-  fp$iota <- NULL
-  
-  return(fp)
-}
-
-
 prepare_logrw <- function(fp, tsEpidemicStart=fp$ss$time_epi_start+0.5){
 
   fp$tsEpidemicStart <- fp$proj.steps[which.min(abs(fp$proj.steps - tsEpidemicStart))]
@@ -117,10 +43,22 @@ rlogistic <- function(t, p){
 }
 
 
-prepare_rlogistic_rw <- function(fp, tsEpidemicStart=fp$ss$time_epi_start+0.5, rw_start=fp$rw_start, rw_dk=NULL){
+#' @param fp model parameters object
+#' @param tsEpidemicStart time step at which epidemic is seeded
+#' @param rw_start time when random walk starts
+#' @param rw_nstep_trans number of RW steps for transition to full variance. If NULL, defaults to 5 steps
+#' @param rw_dk distance between random walk knots. If NULL, defaults to about 1 per year
+prepare_rhybrid <- function(fp,
+                            tsEpidemicStart = fp$ss$time_epi_start+0.5,
+                            rw_start = fp$rw_start,
+                            rw_nstep_trans = fp$rw_nstep_trans,
+                            rw_dk = NULL){
 
   if(is.null(rw_start))
     rw_start <- max(fp$proj.steps)
+
+  if(is.null(rw_nstep_trans))
+    rw_nstep_trans <- 5
 
   fp$tsEpidemicStart <- fp$proj.steps[which.min(abs(fp$proj.steps - tsEpidemicStart))]
 
@@ -135,6 +73,8 @@ prepare_rlogistic_rw <- function(fp, tsEpidemicStart=fp$ss$time_epi_start+0.5, r
 
   rt$n_rw <- n_rw
   rt$n_param <- 4+rt$n_rw  # 4 parameters for rlogistic
+
+  rt$rw_nstep_trans <- pmin(rw_nstep_trans, n_rw)
   
   ## Random walk design matrix
   if(!is.null(rw_dk))
@@ -143,11 +83,11 @@ prepare_rlogistic_rw <- function(fp, tsEpidemicStart=fp$ss$time_epi_start+0.5, r
     rt$rw_knots <- seq(min(rt$rw_steps), max(rt$rw_steps), length.out=n_rw+1)
   rt$rwX <- pmin(pmax(outer(rt$rw_steps, rt$rw_knots[1:n_rw], "-"), 0), 1)  # piecewise linear interpolation
 
-  rt$eppmod <- "rlogistic_rw"
+  rt$eppmod <- "rhybrid"
   fp$rt <- rt
 
   if(!exists("eppmod", fp))
-    fp$eppmod <- "rlogistic_rw"
+    fp$eppmod <- "rhybrid"
   fp$iota <- NULL
   
   return(fp)
@@ -155,19 +95,24 @@ prepare_rlogistic_rw <- function(fp, tsEpidemicStart=fp$ss$time_epi_start+0.5, r
 
 
 create_rvec <- function(theta, rt){
-  if(rt$eppmod == "rlogistic_rw"){
+  if(rt$eppmod == "rhybrid"){
     par <- theta[1:4]
     par[3] <- exp(par[3])
     rvec <- rlogistic(rt$rlogistic_steps, par)
-    rvec <- c(rvec, rvec[length(rt$rlogistic_steps)] + rt$rwX %*% theta[4+1:rt$n_rw])
+
+    th_rw <- theta[4+1:rt$n_rw]
+
+    if(rt$rw_nstep_trans > 0){
+      xx <- seq_len(rt$rw_nstep_trans)
+      th_rw[xx] <- th_rw[xx] * xx / (rt$rw_nstep_trans + 1L)
+    }
+    
+    rvec <- c(rvec, rvec[length(rt$rlogistic_steps)] + rt$rwX %*% th_rw)
+
     return(exp(rvec))
   }
-  if(rt$eppmod == "rhybrid"){
-    u <- theta[1:rt$n_splines]
-    rvec <- log(rt$splineX %*% u)
-    rvec <- c(rep(0, rt$nsteps_preepi), rvec, rvec[length(rvec)] + rt$rwX %*% theta[rt$n_splines+1:rt$n_rw])
-    return(exp(rvec))
-  }
+  else
+    stop(paste(rt$eppmod, "is not impmented in create_rvec()"))
 }
 
 
@@ -191,32 +136,22 @@ extend_projection <- function(fit, proj_years){
   fpnew$SIM_YEARS <- as.integer(proj_years)
   fpnew$proj.steps <- with(fpnew$ss, seq(proj_start+0.5, proj_start-1+fpnew$SIM_YEARS+0.5, by=1/hiv_steps_per_year))
 
-  if(fp$eppmod == "rlogistic_rw"){
+  if(fp$eppmod == "rhybrid"){
     idx1 <- 5  # start of random walk parameters
     idx2 <- 4+fp$rt$n_rw
-    fpnew <- prepare_rlogistic_rw(fpnew, rw_dk=diff(fp$rt$rw_knots[1:2]))
-  } else if(fp$eppmod == "rhybrid"){
-    idx1 <- fp$rt$n_splines+1L # start of random walk parameters
-    idx2 <- fp$rt$n_splines+fp$rt$n_rw
-    fpnew <- prepare_hybrid_r(fpnew, rw_dk=diff(fp$rt$rw_knots[1:2]))
+    fpnew <- prepare_rhybrid(fpnew, rw_dk=diff(fp$rt$rw_knots[1:2]))
   } else if(fp$eppmod == "logrw") {
     idx1 <- 1L
     idx2 <- fp$rt$n_rw
     fpnew <- prepare_logrw(fpnew)
   }
 
-  if(exists("prior_args", fp) && exists("rw_prior_shape", fp$prior_args))
-    sh <- fp$prior_args$rw_prior_shape
-  else
-    sh <- eppasm::rw_prior_shape
-  
-  if(exists("prior_args", fp) && exists("rw_prior_rate", fp$prior_args))
-    rate <- fp$prior_args$rw_prior_rate
-  else
-    rate <- eppasm::rw_prior_rate
-  
   theta <- fit$resample[,idx1:idx2, drop=FALSE]
-  fit$rw_sigma <- sqrt(sample_invgamma_post(theta, sh, rate))
+
+  if(!is.null(fp$prior_args$rw_prior_sd))
+     fit$rw_sigma <- fp$prior_args$rw_prior_sd
+  else
+    fit$rw_sigma <- rw_prior_sd
   
   nsteps <- fpnew$rt$n_rw - fp$rt$n_rw